kafka instance api access utils creation (separated from kafka instance api utils)
