+++ src/test/java/io/managed/services/test/kafka/KafkaAccessControlTest.java
@@ -20,6 +20,7 @@ import io.managed.services.test.client.kafka.KafkaAuthMethod;
 import io.managed.services.test.client.kafka.KafkaConsumerClient;
 import io.managed.services.test.client.kafka.KafkaProducerClient;
 import io.managed.services.test.client.kafkainstance.KafkaInstanceApi;
+import io.managed.services.test.client.kafkainstance.KafkaInstanceApiAccessUtils;
 import io.managed.services.test.client.kafkainstance.KafkaInstanceApiUtils;
 import io.managed.services.test.client.kafkamgmt.KafkaMgmtApiUtils;
 import io.managed.services.test.client.oauth.KeycloakLoginSession;
@@ -64,7 +65,6 @@ import static org.testng.Assert.assertTrue;
  * </ul>
  */
 public class KafkaAccessControlTest extends TestBase {
-    // TODO At least all described tests in the document are automated
     //TODO KafkaMgmtAPIPermissionTest (migrate completely)
     //TODO KafkaAdminPermissionTest (migrate completely)
     //TODO KafkaInstanceAPITest (migrate all permission tests)
@@ -86,6 +86,7 @@ public class KafkaAccessControlTest extends TestBase {
     private ApplicationServicesApi adminAPI;
 
     private ServiceAccount primaryServiceAccount;
+    private ServiceAccount secondaryServiceAccount;
 
     private KafkaRequest kafka;
     private KafkaInstanceApi kafkaInstanceAPIPrimaryUser;
@@ -130,7 +131,7 @@ public class KafkaAccessControlTest extends TestBase {
 
         primaryServiceAccount =
                 SecurityMgmtAPIUtils.applyServiceAccount(primaryAPI.securityMgmt(), SERVICE_ACCOUNT_NAME);
-        primaryServiceAccount =
+        secondaryServiceAccount =
                 SecurityMgmtAPIUtils.applyServiceAccount(primaryAPI.securityMgmt(), DEFAULT_SERVICE_ACCOUNT_NAME);
 
         // create the kafka admin
@@ -158,7 +159,7 @@ public class KafkaAccessControlTest extends TestBase {
                 Environment.ADMIN_PASSWORD));
 
         // get default ACLs for Kafka Instance
-        defaultPermissionsList = KafkaInstanceApiUtils.getDefaultACLs(kafkaInstanceAPIPrimaryUser);
+        defaultPermissionsList = KafkaInstanceApiAccessUtils.getDefaultACLs(kafkaInstanceAPIPrimaryUser);
 
         // create topic that is needed to perform some permission test (e.g., messages consumption)
         KafkaInstanceApiUtils.applyTopic(kafkaInstanceAPIPrimaryUser, TOPIC_NAME_EXISTING_TOPIC);
@@ -173,9 +174,9 @@ public class KafkaAccessControlTest extends TestBase {
             primaryAdmin.close();
         }
 
-        //Clear all but default ACLs....
+        //Clear all but default ACLs (i.e., those that instance had at moment of creation/application)
         try {
-            KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+            KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
         } catch (Throwable t) {
             LOGGER.error("clean extra ACLs error: ", t);
         }
@@ -291,7 +292,7 @@ public class KafkaAccessControlTest extends TestBase {
     public void testACLsDefaultServiceAccountCanListTopic() {
 
         // removal of possibly existing additional ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         LOGGER.info("Test default service account ability to list topics");
         primaryAdmin.listTopics();
@@ -302,7 +303,7 @@ public class KafkaAccessControlTest extends TestBase {
     public void testACLsDefaultServiceAccountCannotProduceAndConsumeMessages() {
 
         // removal of possibly existing additional ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         LOGGER.info("Test default service account inability to produce and consume data from topic {}", TOPIC_NAME_EXISTING_TOPIC);
         assertThrows(GroupAuthorizationException.class, () -> bwait(testTopic(
@@ -322,7 +323,7 @@ public class KafkaAccessControlTest extends TestBase {
     public void testACLsDefaultServiceAccountCannotCreateACLs() {
 
         // removal of possibly existing additional ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
         LOGGER.info("Test default service account inability to create ACL");
         assertThrows(ClusterAuthorizationException.class, () -> primaryAdmin.addAclResource(ResourceType.TOPIC));
     }
@@ -334,8 +335,8 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test ability of default service account with additional ACLs to create topic");
 
         // clean and add ACLs on all resources for default account
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC)
         );
 
@@ -359,8 +360,8 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test ability of default service account with additional ACLs to create topic");
 
         // removal of possibly existing additional ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC, AclResourceType.GROUP, AclResourceType.TRANSACTIONAL_ID)
         );
 
@@ -384,8 +385,8 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test ability of default service account with additional ACLs to list consumer groups");
         // removal of possibly existing additional ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC, AclResourceType.GROUP, AclResourceType.TRANSACTIONAL_ID)
         );
 
@@ -401,8 +402,8 @@ public class KafkaAccessControlTest extends TestBase {
         final String groupId = "cg-3";
 
         // add ACLs on all resources for default account
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC, AclResourceType.GROUP, AclResourceType.TRANSACTIONAL_ID)
         );
 
@@ -431,8 +432,8 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test ability of default service account with additional ACLs to list consumer groups");
 
         // add ACLs on all resources for default account
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC, AclResourceType.GROUP, AclResourceType.TRANSACTIONAL_ID)
         );
 
@@ -467,7 +468,7 @@ public class KafkaAccessControlTest extends TestBase {
 
         // deny rights
         LOGGER.info("new ACL that deny right to read Topics for tested service account is to be applied");
-        var principal = KafkaInstanceApiUtils.toPrincipal(primaryServiceAccount.getClientId());
+        var principal = KafkaInstanceApiAccessUtils.toPrincipal(primaryServiceAccount.getClientId());
         var acl = new AclBinding()
                 .principal(principal)
                 .resourceType(AclResourceType.TOPIC)
@@ -503,8 +504,8 @@ public class KafkaAccessControlTest extends TestBase {
         final String topicNamePrefixed = prefix + topicName;
 
         // add ACLs on all resources for default account
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.TOPIC)
         );
         // add ACL to Deny deletion of topic with prefix for all of users
@@ -545,14 +546,14 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test ACL ability to deny rights to work with consumer group (i.e., read )");
 
         // add ACLs on all resources for default account
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
-        KafkaInstanceApiUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.applyAllowAllACLsOnResources(kafkaInstanceAPIPrimaryUser, primaryServiceAccount,
                 List.of(AclResourceType.GROUP, AclResourceType.TOPIC)
         );
 
         // add ACL: Deny Topic Describe to all
         LOGGER.info("ACL: deny describe Topic to service account");
-        var principal = KafkaInstanceApiUtils.toPrincipal(primaryServiceAccount.getClientId());
+        var principal = KafkaInstanceApiAccessUtils.toPrincipal(primaryServiceAccount.getClientId());
         var acl = new AclBinding()
                 // for every user
                 .principal(principal)
@@ -605,7 +606,7 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test ACL default setting, secondary user can list ACLs");
 
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         kafkaInstanceAPISecondaryUser.getAcls(null, null, null, null, null, null, null, null, null, null);
     }
@@ -616,7 +617,7 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test ACL default setting, secondary user cannot delete ACLs");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         assertThrows(ApiForbiddenException.class, () ->
                 kafkaInstanceAPISecondaryUser.deleteAcls(null, null, null, null, null, null));
@@ -627,15 +628,14 @@ public class KafkaAccessControlTest extends TestBase {
     public void testACLsTopicsAllowAllForSecondaryUser() {
 
         LOGGER.info("Test default setting, secondary user cannot delete ACLs");
-        final String topicName = "topic-name-acl-secondary-acc";
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
-        assertThrows(AuthorizationException.class, () -> primaryAdmin.deleteAclResource(ResourceType.ANY));
+        final String topicName = "topic-name-acl-secondary-acc";
 
         // add ACL to Deny deletion of topic with prefix for all of users
         var acl = new AclBinding()
-                .principal(KafkaInstanceApiUtils.toPrincipal(Environment.SECONDARY_USERNAME))
+                .principal(KafkaInstanceApiAccessUtils.toPrincipal(Environment.SECONDARY_USERNAME))
                 .resourceType(AclResourceType.TOPIC)
                 .patternType(AclPatternType.LITERAL)
                 .resourceName("*")
@@ -662,14 +662,14 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test ACL default setting, secondary user cannot delete ACLs");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         // add ACL to Deny deletion of topic with prefix for all of users
         var acl = new AclBinding()
-                .principal(KafkaInstanceApiUtils.toPrincipal(Environment.SECONDARY_USERNAME))
+                .principal(KafkaInstanceApiAccessUtils.toPrincipal(Environment.SECONDARY_USERNAME))
                 .resourceType(AclResourceType.TOPIC)
                 .patternType(AclPatternType.LITERAL)
-                .resourceName("xxxx")
+                .resourceName("xyz")
                 .permission(AclPermissionType.ALLOW)
                 .operation(AclOperation.ALL);
 
@@ -682,19 +682,25 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test ACL default setting, secondary user cannot delete ACLs");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         assertThrows(ApiForbiddenException.class, () -> kafkaInstanceAPIAdminUser.deleteAcls(
             AclResourceTypeFilter.TOPIC, "xx", null, null, null, null));
     }
-    //@Ignore
-    @SneakyThrows
+    // TODO delete and change the other one to testAdminUserCanChangeTheKafkaInstanceOwner
     @Test (priority = 1)
     public void testAdminUserCanChangeTheKafkaInstanceOwner() {
 
+    }
+
+    @Ignore
+    @SneakyThrows
+    @Test (priority = 1)
+    public void testAdminUserCanChangeTheKafkaInstanceOwnerr() {
+
         LOGGER.info("Test switch owner (admin) of kafka instance from primary user to secondary");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPIPrimaryUser, defaultPermissionsList);
 
         var authorizationTopicName = "authorization-topic-name";
 
@@ -723,11 +729,11 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test ACL default setting, new Instance owner (Secondary User) can create ACLs");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
 
         // add ACL to Deny deletion of topic with prefix for all of users
         var acl = new AclBinding()
-                .principal(KafkaInstanceApiUtils.toPrincipal(Environment.PRIMARY_USERNAME))
+                .principal(KafkaInstanceApiAccessUtils.toPrincipal(Environment.PRIMARY_USERNAME))
                 .resourceType(AclResourceType.TOPIC)
                 .patternType(AclPatternType.LITERAL)
                 .resourceName("*")
@@ -743,7 +749,7 @@ public class KafkaAccessControlTest extends TestBase {
 
         LOGGER.info("Test  new Instance owner (Secondary User) can create ACLs");
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
 
         kafkaInstanceAPISecondaryUser.deleteAcls(
             AclResourceTypeFilter.TOPIC, "xx", null, null, null, null);
@@ -756,39 +762,31 @@ public class KafkaAccessControlTest extends TestBase {
         LOGGER.info("Test old instance owner (primary user) can list ACLs");
 
         // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
+        KafkaInstanceApiAccessUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
 
         kafkaInstanceAPIPrimaryUser.getAcls(null, null, null, null, null, null, null, null, null, null);
     }
 
-    @Test(priority = 1, dependsOnMethods = "testAdminUserCanChangeTheKafkaInstanceOwner")
-    @SneakyThrows
-    public void testOldInstanceOwnerCannotDeleteACLs() {
 
-        LOGGER.info("Test old instance owner (primary user) can list ACLs");
-        // clean ACLs
-        KafkaInstanceApiUtils.removeAllButDefaultACLs(kafkaInstanceAPISecondaryUser, defaultPermissionsList);
-
-        assertThrows(ApiForbiddenException.class, () ->
-                kafkaInstanceAPIPrimaryUser.deleteAcls(null, null, null, null, null, null));
-    }
-
+    // TODO unignore
     @Ignore
-    //TODO former owner can
     @Test(priority = 1, dependsOnMethods = "testAdminUserCanChangeTheKafkaInstanceOwner")
     public void testFormerOwnerCanNotDeleteTheKafkaInstance() {
+        LOGGER.info("Test old instance owner (primary user) can list ACLs");
         // should failKafkaControlManagerAPIPermissionsTestKafkaControlManagerAPIPermissionsTest
         assertThrows(ApiNotFoundException.class, () -> primaryAPI.kafkaMgmt().deleteKafkaById(kafka.getId(), true));
     }
 
+    @Ignore
     @Test(priority = 1)
     public void testAlienUserCanNotDeleteTheKafkaInstance() {
         // should fail
         assertThrows(ApiNotFoundException.class, () -> alienAPI.kafkaMgmt().deleteKafkaById(kafka.getId(), true));
     }
 
+    @Ignore
     @SneakyThrows
-    @Test (priority = 2)
+    @Test (priority = 2, dependsOnMethods = "testAdminUserCanChangeTheKafkaInstanceOwner")
     // test is should be executed as last one.
     public void testNewOwnerCanDeleteTheKafkaInstance() {
         KafkaMgmtApiUtils.cleanKafkaInstance(secondaryAPI.kafkaMgmt(), KAFKA_INSTANCE_NAME);
